Moved configuration in separate package
